pyproject.toml
[build-system]
requires = [
  "setuptools >= 61.0.0",
  "wheel",
  "setuptools_scm[toml] >= 6.2"
]
build-backend = "setuptools.build_meta"

[project]
name = "junigraph"
description = "Create SVG containing a visual representation of a junifer pipeline."
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE.md"}
authors = [
    {email = "l.sasse@fz-juelich.de"},
    {name = "Leonard Sasse"},
]

keywords = ["junifer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Natural Language :: English",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "PyYAML",
    "svgwrite",
]
dynamic = ["version"]

[project.urls]
repository = "https://github.com/LeSasse/junigraph"

[project.scripts]
junigraph = "junigraph.main:main"

#[project.optional-dependencies]
#dev = ["tox"]

################
# Tool configs #
################

[tool.setuptools]
packages = ["junigraph"]

[tool.setuptools_scm]
version_scheme = "python-simplified-semver"
local_scheme = "no-local-version"
write_to = "junigraph/_version.py"

[tool.black]
line-length = 79
target-version = ["py37"]